+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 
 ## [2.9.1] - 2022-05-18
 ### Added
+- gen_class.TimeFormat: Class to store times in pre-defined or user-defined time formats and retrieve with the same time setting. 
 - gen_class.ArgParser.delete_arg: Deletes a value from the args_array attribute.
 - gen_class.ArgParser.update_arg: Updates a value in the args_array attribute.
 - gen_class.ArgParser.insert_arg: Inserts a key and value into the args_array attribute.